+++ crates/block2/src/debug.rs
@@ -7,7 +7,7 @@ use std::ffi::CStr;
 use crate::{abi, ffi, Block, ConcreteBlock, GlobalBlock, RcBlock};
 
 #[derive(Clone, Copy, PartialEq, Eq)]
-struct Isa(*const abi::Class);
+struct Isa(*const ffi::Class);
 
 impl Isa {
     fn is_global(self) -> bool {
@@ -222,7 +222,7 @@ mod tests {
         let isa = Isa(unsafe { &ffi::_NSConcreteStackBlock });
         assert!(!isa.is_global());
         assert!(isa.is_stack());
-        let isa = Isa(unsafe { &abi::_NSConcreteMallocBlock });
+        let isa = Isa(unsafe { &ffi::private::_NSConcreteMallocBlock });
         assert!(!isa.is_global());
         assert!(!isa.is_stack());
         let isa = Isa(ptr::null());